Move to using the cognito hosted ui.  Remove the login and signup containes
